Rewrote ChatItemRewriter in 1.12 -> 1.11.2 to use MCStructs
